+++ tree_shaking/__main__.py
@@ -1,7 +1,8 @@
 from argsense import cli
+
 from .export import dump_tree
-from .graph import build_module_graphs
 from .graph import build_module_graph
+from .graph import build_module_graphs
 
 cli.add_cmd(build_module_graphs)
 cli.add_cmd(build_module_graph)
@@ -17,9 +18,9 @@ if __name__ == '__main__':
     #       prepare: make sure `chore/site_packages` latest:
     #           pox sidework/merge_external_venv_to_local_pypi.py .
     #           pox build/init.py make-site-packages --remove-exists
-    
+
     # pox -m tree_shaking build-module-graphs demo_config/modules.yaml
-    
+
     # pox -m tree_shaking dump-tree <file_i> <dir_o>
     # pox -m tree_shaking dump-tree <file_i> <dir_o> --copyfiles
     cli.run()