+++ tree_shaking/file_parser.py
@@ -3,10 +3,9 @@ import atexit
 import typing as t
 from contextlib import contextmanager
 
-from lk_logger import parallel_printing
 from lk_utils import fs
-from lk_utils import p
 
+from .cache import cache_root
 from .cache import file_cache
 from .module import ModuleInfo
 from .module import ModuleInspector
@@ -16,7 +15,7 @@ from .module import T as T0
 from .path_scope import path_scope
 
 module_inspector = ModuleInspector(
-    ignores=fs.load(p('_cache/ignores.txt')).splitlines()
+    ignores=fs.load('{}/ignores.txt'.format(cache_root)).splitlines()
 )
 _broken = set()
 
@@ -28,13 +27,11 @@ class T(T0):
     Node = t.Union[ast.Import, ast.ImportFrom]
 
 
-# noinspection PyMethodMayBeStatic
 class FileParser:
-    
     def __init__(self, file: T.FilePath) -> None:
         self.file = file
         self.dir = fs.parent(file)
-        
+
         if self.file in path_scope.path_2_module:
             self.base_dir = self.dir
             self.base_module_segs = ()
@@ -46,14 +43,15 @@ class FileParser:
                 if self.file.startswith(top_path + '/'):
                     self.base_dir = top_path
                     self.base_module_segs = (
-                        top_name, *self.dir[len(top_path) + 1:].split('/')
+                        top_name,
+                        *self.dir[len(top_path) + 1 :].split('/'),
                     )
                     break
             else:
                 raise Exception(
                     'file should be existed in registered path scopes', file
                 )
-    
+
     @property
     def module_info(self) -> ModuleInfo:
         if self.base_module_segs:
@@ -70,9 +68,9 @@ class FileParser:
             base_dir=self.base_dir,
             full_name=module_name,
         )
-    
+
     def parse_imports(self) -> T.ImportsInfo:
-        # print(':dv2sp', 'start', self.file)
+        # print(':dv2p', 'start', self.file)
         for node, line in file_cache.parse_nodes(self.file):
             for module in self._get_module_info(node, line):
                 try:
@@ -82,7 +80,7 @@ class FileParser:
                         _broken.add(module.id)
                         # with _err_records.recording():
                         #     print(
-                        #         ':v3l',
+                        #         ':v5l',
                         #         '{}:{}'.format(self.file, node.lineno),
                         #         line.strip(),
                         #         module,
@@ -96,11 +94,12 @@ class FileParser:
                     continue
                 else:
                     yield module, path
-        # print(':vsp', 'end', self.file)
-    
+        # print(':vp', 'end', self.file)
+
     def _check_if_relative_import(self, line: str) -> int:
         x = line.lstrip().split()[1]
-        if not x: raise Exception('empty import', self.file, line)
+        if not x:
+            raise Exception('empty import', self.file, line)
         dot_cnt = 0
         for ch in x:
             if ch == '.':
@@ -108,7 +107,7 @@ class FileParser:
             else:
                 break
         return dot_cnt
-    
+
     def _get_module_info(
         self, node: T.Node, line: str
     ) -> t.Iterator[T.ModuleInfo]:  # noqa
@@ -116,7 +115,7 @@ class FileParser:
             if dot_cnt == 1:
                 base_module = '.'.join(self.base_module_segs)
             else:
-                base_module = '.'.join(self.base_module_segs[:-(dot_cnt - 1)])
+                base_module = '.'.join(self.base_module_segs[: -(dot_cnt - 1)])
         else:
             base_module = None
         if base_module:
@@ -125,7 +124,7 @@ class FileParser:
             )
         else:
             base_dir = None
-        
+
         if isinstance(node, ast.Import):
             for alias in node.names:
                 if base_module:
@@ -152,7 +151,9 @@ class FileParser:
                     else:
                         print(
                             ':v4l',
-                            self.file, node.lineno, line,
+                            self.file,
+                            node.lineno,
+                            line,
                             {
                                 k: getattr(node, k)
                                 for k in dir(node)
@@ -173,32 +174,33 @@ class FileParser:
                     level=dot_cnt,
                     base_dir=base_dir,
                 )
-    
+
     def _get_module_path(self, module: T.ModuleInfo) -> T.FilePath:
         # assert '//' not in module_inspector.find_module_path(module), module
         return module_inspector.find_module_path(module)
 
 
 class ErrorRecords:
-    
     def __init__(self) -> None:
         self._records = []
         atexit.register(self.save)
-    
+
     @contextmanager
-    def recording(self) -> t.Iterator:
-        with parallel_printing(self._log):
-            yield
-    
+    def recording(self) -> t.Generator:
+        # FIXME: with parallel_printing(self._log):
+        yield
+
     def _log(self, msg: str) -> None:
         self._records.append(str(msg))
-    
+
     def save(self) -> bool:
         if self._records:
-            fs.dump(self._records, p('_cache/errors.txt'))
+            fs.dump(self._records, '{}/errors.txt'.format(cache_root))
             print(
-                'found {} errors. see log at "_cache/errors.txt"'
-                .format(len(self._records)), ':v8s'
+                'found {} errors. see log at "{}/errors.txt"'.format(
+                    len(self._records), cache_root
+                ),
+                ':v8s',
             )
             return True
         return False