+++ tree_shaking/patch.py
@@ -4,10 +4,9 @@ from lk_utils import fs
 
 
 class Patch:
-    
-    def __init__(self) -> None:
-        cfg = fs.load(fs.xpath('patches/implicit_import_hooks.yaml'))
-        
+    def __init__(self, hooks_file: str) -> None:
+        hooks = fs.load(hooks_file)
+
         # {
         #   module: {
         #       'imports': (relpath, ...),
@@ -19,19 +18,20 @@ class Patch:
         #   }, ...
         # }
         self._patches = {}
-        for k, v in cfg.items():
+        for k, v in hooks.items():
             self._patches[k] = {
                 'files': tuple(v.get('files', ())),
-                'imports': tuple(v.get('imports', ()))
+                'imports': tuple(v.get('imports', ())),
             }
-    
+
     def __contains__(self, module_name: str) -> bool:
         return module_name in self._patches
-    
-    def __getitem__(self, module_name: str) -> t.TypedDict('PatchItem', {
-        'files'  : t.Tuple[str, ...], 'imports': t.Tuple[str, ...]
-    }):
+
+    def __getitem__(self, module_name: str) -> t.TypedDict(
+        'PatchItem', {'files': t.Tuple[str, ...], 'imports': t.Tuple[str, ...]}
+    ):
         return self._patches[module_name]
 
 
-patch = Patch()
+implicit_hooks_file = fs.here('patches/implicit_import_hooks.yaml')
+patch = Patch(implicit_hooks_file)